Set up ResendConfirmation React component
