+++ projects/cloud/app/views/devise/sessions/new.html.erb
@@ -1,5 +1,3 @@
-<h2>Log in</h2>
-
 <%= javascript_pack_tag 'application' %>
 <link
   rel="stylesheet"
